sqrt

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -13,7 +13,7 @@ int _sqrt_recursion(int n)
 	return (helper(n, 0, n));
 }
 /**
- * helper function that uses binary search to find sqrt
+ * helper - function that uses binary search to find sqrt
  * @n: input
  * @i: start of search
  * @j: end of search
@@ -21,25 +21,25 @@ int _sqrt_recursion(int n)
  */
 int helper(int n, int i, int j)
 {
-        int mid;
-	int square;
-	
+	int mid;
+	long int square;
+
 	if (i > j)
-        {
-                return (-1);
-        }
+	{
+		return (-1);
+	}
 	mid = i + (j - i) / 2;
-        square = mid * mid;
+	square = mid * mid;
 	if (n == square)
-        {
-                return (mid);
-        }
-        else if (square < n)
-        {
-                return (helper(n, mid + 1, j));
-        }
-        else
-        {
-                return (helper(n, i, mid + 1));
-        }
+	{
+		return (mid);
+	}
+	else if (square < n)
+	{
+		return (helper(n, mid + 1, j));
+	}
+	else
+	{
+		return (helper(n, i, mid + 1));
+	}
 }
